ref(lineage): move extract_lineage_chain to lineage_utils
